Monday work

+++ Dragon 2/src/org/usfirst/frc/team6201/robot/subsystems/DriveTrain.java
@@ -9,6 +9,7 @@ import com.ctre.phoenix.motorcontrol.NeutralMode;
 import com.ctre.phoenix.motorcontrol.can.WPI_TalonSRX;
 
 import edu.wpi.first.wpilibj.ADXRS450_Gyro;
+import edu.wpi.first.wpilibj.DriverStation;
 import edu.wpi.first.wpilibj.command.Subsystem;
 
 /**
@@ -162,7 +163,7 @@ public class DriveTrain extends Subsystem {
     	
     	Robot.dt.left1.getSensorCollection().setPulseWidthPosition(0, 0);
     	Robot.dt.right1.getSensorCollection().setPulseWidthPosition(0, 0);
-    	
+    	//DriverStation.reportWarning("TEsting", false);
     }
     
     /**
@@ -174,7 +175,7 @@ public class DriveTrain extends Subsystem {
 	 */
     public double getDistanceTraveled() {
     	
-    	return -(Robot.dt.left1.getSensorCollection().getPulseWidthPosition() / 4096) * 6 * Math.PI;
+    	return -((double) Robot.dt.left1.getSensorCollection().getPulseWidthPosition() / 4096.0) * 6 * Math.PI;
     	
     }
